requirements.txt
llama-cpp-python
llama-cpp-python[server]
llama-index
fastapi
uvicorn
sentence-transformers
langchain